requirements.txt
# requirements.txt

# For data handling and numerical operations
pandas
numpy

# For plotting the backtest results
matplotlib

# For loading API keys from the .env file
python-dotenv

# For fetching data from Polygon.io
polygon-api-client

# For the S&P 500 scraping (used by pandas)
requests
lxml

# For sentiment analysis with Gemini, DeepSeek
google-generativeai
openai